React use common DeleteButton

+++ src/react/components/buttons/DeleteButton.jsx
@@ -2,16 +2,13 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import { Message, Confirm, Button,Tooltip } from 'oskari-ui';
-
+import { IconButton } from 'oskari-ui/components/buttons';
 import { SecondaryButton } from './';
-import { DeleteOutlined } from '@ant-design/icons';
-import { red } from '@ant-design/colors'
+import { DeleteIcon } from '../icons'
 
-const DeleteIcon = styled(DeleteOutlined)`
-    color: ${red.primary};
-    cursor: pointer;
+const StyledButton = styled(Button)`
+    font-size: 16px;
 `;
-
 
 const getButton = (type, disabled ) => {
     const props = {
@@ -19,23 +16,21 @@ const getButton = (type, disabled ) => {
         disabled
     }
     if (type === 'icon') {
-        return <DeleteIcon {...props}/>;
+        return <IconButton icon={<DeleteIcon/>} {...props}/>;
     }
     if (type === 'button') {
-        return <Button  {...props}><DeleteIcon/></Button>;
+        return <StyledButton  {...props}><DeleteIcon/></StyledButton>;
     }
-    return <SecondaryButton {...props} danger type="delete"/>;
+    if (type === 'label') {
+        return <SecondaryButton {...props} danger type="delete"/>;
+    }
 };
-const getMsg = key => <Message messageKey={key} bundleKey='oskariui'/>
 
-
-// TODO: update bundles to use type and remove icon
 export const DeleteButton = ({ 
     onConfirm,
-    icon,
-    type = icon ? 'icon' : '',
+    type,
     title = <Message messageKey='messages.confirmDelete' bundleKey='oskariui'/>,
-    tooltip = getMsg('buttons.delete'),
+    tooltip = <Message messageKey={'buttons.delete'} bundleKey='oskariui'/>,
     disabled = false
 }) => {
     const placement = tooltip ? 'bottom' : 'top';
@@ -60,8 +55,7 @@ export const DeleteButton = ({
 DeleteButton.propTypes = {
     onConfirm: PropTypes.func.isRequired,
     title: PropTypes.node,
-    icon: PropTypes.bool,
     tooltip: PropTypes.node,
     disabled: PropTypes.bool,
-    type: PropTypes.oneOf(['icon', 'button', 'text'])
+    type: PropTypes.oneOf(['icon', 'button', 'label']).isRequired
 };
